Commented Move to Player out in the basic enemy object to make 2 different Types (melee and ranged)
Melee has the same Movement Logic like before. Ranged Enemy Movement has some more Code now. We set a min and max distance and standard movespeed. We save the distance and direction to the player. Now we can seperate 3 different Ways to act. 1. if the enemy is too close to the player, he backs off. 2. If the enemy is too far from the player he moves towards him and Last if he´s in the radius between min and max the enemy kites to the left and right and moves kinda in a circle.